requirements.txt
pillow==11.3.0
pillow_heif==1.1.0
tqdm==4.67.1
